Refactor pregel into algo
